Module lignin_schema::aspects

👎 Deprecated:
Don't specify aspects directly, which may be too limiting. Generic type parameter resolution should never fail as there are no overloaded implementations.

Structs

Attribute	Deprecated

Content	Deprecated

Event	Deprecated