Module lignin_schema::html::elements

See https://developer.mozilla.org/en-US/docs/Web/HTML/Element.

Structs

a

abbr

acronym	Deprecated

address

applet	Deprecated

area
Empty.

article

aside

audio

b

base
Empty.

basefont	Deprecated
Empty.

bdi

bdo

bgsound	Deprecated
Empty.

big	Deprecated

blink	Deprecated

blockquote

body

br
Empty. Produces a line break (carriage-return) in text.

button

canvas

caption

center	Deprecated

cite

code

col
Empty.

colgroup

command	Deprecated
Empty.

content	Deprecated

data

datalist

dd

del

details

dfn

dialog

dir	Deprecated

div
A generic layout container without semantic meaning, by default rendered with display: block.

dl

dt

element	Deprecated

em

embed
Empty.

fieldset

figcaption

figure

font	Deprecated

footer

form

frame	Deprecated
Empty.

frameset	Deprecated

h1

h2

h3

h4

h5

h6

head

header

hgroup

hr
Empty. Thematic break, for example as sibling between <p> elements to mark a change of scene or topic within one containing section.

html

i

iframe

image	Deprecated
Empty.

img
Empty.

input
Empty.

ins

isindex	Deprecated
Empty.

kbd

keygen	Deprecated
Empty.

label

legend

li

link
Empty.

listing	Deprecated

main

map

mark

marquee	Deprecated

menu

menuitem	Deprecated
Empty.

meta
Empty.

meter

multicol	Deprecated

nav

nextid	Deprecated
Empty.

nobr	Deprecated

noembed	Deprecated

noframes	Deprecated

noscript

object

ol

optgroup

option

output

p

param
Empty.

picture

plaintext	Deprecated

pre

progress

q

rb

rp

rt

rtc

ruby

s

samp

script

section

select

shadow	Deprecated

slot

small

source
Empty.

spacer	Deprecated
Empty.

span

strike	Deprecated

strong

style

sub

summary

sup

table

tbody

td

template

textarea

tfoot

th

thead

time

title

tr

track
Empty.

tt	Deprecated

u

ul

var

video

wbr
Empty.

xmp	Deprecated